pylog/requirements.txt
PyYAML
watchdog
deprecated
typing
